Add option for image position in card component
Add option for image position in card component, only show image to the side if screen is big enough. Showing the image to the side when the screen is small looks off.
Create useValidator composable to be used for props that have specific values. Can add more validation criteria in the future as needed.
Change CardComponent to use script setup to be consistent with other files, and use h2 to achieve the same size without modifying h1.
Document the use of BEM naming convention (prop-name--prop-value) for class names of props with specific values.