Give credit to SVG Creators properly
Create LICENSE.md to specify that the website is under CC license, while the SVGs are under MIT license.
Remove license part in README.md as license already has a dedicated file for it.
Use different SVG for navbar x as the license part isn't present and can't be found.
Add comment for the license of hamburger icon.