Add option for image position in card component
Fix a bug where the text appears on top of images for cards with side image positions.
Fix a bug where the images aren't the same size as other cards, an issue for both side and top image positions. (Likely caused by .card-img not being specific enough and thus not targeting the actual image)